"Update Python library requirements and enhance build scripts"
In this commit, conandata.yml was deleted as the Python library requirements are now defined in conanfile.py. Incorporated a stricter check for the minimum required C++ standard and for specific compiler versions for better code compatibility. Moreover, modified the function calls in conanfile.py and test_package/conanfiles.py to use more updated and reliable methods. Adjusted build requirements and export sources to improve the building process. Finally, optimized the packaging process to ensure that all necessary files are included in the correct directories.

Contributes to CURA-10951